Trait zcore_drivers::builder::IoMapper

pub trait IoMapper {
    fn query_or_map(&self, paddr: usize, size: usize) -> Option<usize>;
}

A trait implemented in kernel to translate device physical addresses to virtual addresses.

Required Methods

fn query_or_map(&self, paddr: usize, size: usize) -> Option<usize>

Translate the device physical address to virtual address. If not mapped in the kernel page table, map the region specified by the given size.

If an error accurs during translation or mapping, returns None.