Module zcore_drivers::scheme

The Scheme describe some functions must be implemented for device, there are many Scheme traits in this mod.

If you need to develop a new device, just implement the corresponding trait.

The Scheme trait is suitable for any architecture.

Traits

BlockScheme

DisplayScheme

EventScheme

InputScheme

IrqScheme

NetScheme

Scheme

SchemeUpcast

UartScheme