Trait zcore_drivers::scheme::EventScheme

pub trait EventScheme {
    type Event;

    fn trigger(&self, event: Self::Event);
    fn subscribe(&self, handler: EventHandler<Self::Event>, once: bool);
}

Required Associated Types

type Event

Required Methods

fn trigger(&self, event: Self::Event)

Trigger the event manually and call its handler immediately.

fn subscribe(&self, handler: EventHandler<Self::Event>, once: bool)

Subscribe events, call the handler when an input event occurs. If once is ture, unsubscribe automatically after handling.